[rx] switch tests to use rx2. switch implementation away from rx adapter into only extension wrappers.
